python_web_app/requirements.txt
flask==2.3.3
requests==2.31.0
snowflake-connector-python==3.4.0
python-dotenv==1.0.0